protocol: align cross-language wire format and signing layout
Closes the wire-compatibility findings from the 2026-05-02 audit. After this
commit, every implementation produces byte-identical MeshPacket bytes for the
same input, and signature verification works across language boundaries.

Wire format
- C# PacketSerializer: write Guid in RFC 4122 big-endian (was .NET's mixed-
  endian default — broke RREQ dedup hashing the moment routing shipped).
- Swift / C: TTL widened from UInt8 / uint8_t to Int32 (was silently
  truncating values > 255).
- Go SosPriority retyped from byte (couldn't hold 999) to int32; canonical
  on-wire value is 255 anyway, but the constant must be representable.

Signing
- C# PacketSigningService.BuildSignableData rewritten: little-endian + 4-byte
  Type / Priority (was big-endian via Array.Reverse + 1-byte Type, which
  broke verification against every other language).
- Nonce dedup keyed by (source, nonce) tuple (was hex(nonce) only — collision
  surface across senders).

Other security primitives
- Python _zero_memory now actually zeroes (was a no-op despite the name).
- Kotlin Ed25519Service: imported java.security.SecureRandom (was
  org.bouncycastle.crypto.SecureRandom which doesn't exist).
- Rust main.rs disabled — its security/* dependencies don't compile and the
  demo is unrelated to the protocol library itself.

Constants
- SosPriority canonicalised to 255 across all 8 languages (was 999 in some).

+++ c/include/aether/protocol.h
@@ -80,7 +80,7 @@ typedef struct {
     int64_t  timestamp_ms;                             // 8 bytes
     uint8_t  protocol_version;                         // 1 byte
     uint8_t  type;                                     // 1 byte
-    uint8_t  ttl;                                      // 1 byte
+    int32_t  ttl;                                      // 4 bytes (wire is little-endian int32; was uint8_t — truncation bug fixed 2026-05-02)
     uint8_t  priority;                                 // 1 byte
 
     // Variable-length fields (pointers + lengths)
